chore: Update Astro configuration and styles
This commit updates the Astro configuration file (`astro.config.mjs`) to include the `output` option set to "static". It also removes the import of the `globals.css` file in the `Layout.astro` file and adds a preload link for the stylesheet. These changes were made to improve the build process and optimize the loading of styles in the application.